+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Error.1 is a struct instead of a tuple.
 - Error is now a struct.
 - The declarations order is more flexible.
+- Way better error reporting when there is a syntax error in the macro call.
 
 # 0.5.0